승,패,무, player플레이어 클래스로 이전

+++ pro11/src/game/record/Record.java
@@ -45,6 +45,6 @@ public class Record {
 	
 	public double getWinRate() {
 		// 승률 정보를 반환한다.
-		return score[0] / getRecordCount( ) * 100;
+		return (double)score[0] / getRecordCount( ) * 100;
 	}
 }
